+++ resources/js/utils.js
@@ -5,3 +5,18 @@ export function isEmpty(obj) {
       return !obj || !Object.keys(obj).length;
     }
 }
+
+// ステータスコードが200番代か判定する
+export function successfulStatus(code) {
+    if(code / 100 == 2) {
+        return true;
+    }
+    return false;
+}
+
+// オブジェクトを受け取って、GETリクエストのクエリパラメーターに変換する
+export function convertQuery(obj) {
+    return Object.keys(body).map((key) => {
+        return key + "=" + body[key];
+    }).join('&')
+}
